Scan ancestry layers in reverse order
Some feature/namespace detectors requires to process ancestry layers in
reverse order (base layer needs to be processed last). This is due to
fact that some base layers don't contain enough information to
provide correct namespaces and a namespaces needs to be inherited from
previous layer. Layers need to be scanned sequentially to be able
inherit namespaces in between layers.