Store layer info outside parallel go routines
Current Clair implementation process layers in go-routines. This doesn't
allow sharing any information in between layers. This commit update
go-routines to only fetch layer, extract it and detect content. Storing
layer information to database is moved outside of go-routines.

This allow sharing an information (namespace) in between layers in case
namespace is not defined in a layer.

This is just preparation commit and other commits which implement
sharing information in between layers will follow.