Add code for a new Photon OS driver
Currently, Photon OS is not supported by Clair and it will a valuable
addition for all Photon users if that's changed.

This code is not created for a particular version of Photon but
instead fetches information about the released versions of
Photon. If, a new version of Photon is released Clair will
automatically support it.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>