Ensure Amazon Linux is correctly identified
Amazon Linux-based images are non-deterministically failing to be
detected with the correct namespace name. The centos regex in the
redhatrelease detector is too permissive causing Amazon Linux AMI
based images to return a namespace name of "AMI". Additionally, Amazon
Linux 2 is detected to have the namespace name of "amazon". The correct
namespace name is "amzn".

This change adds a more restrictive check for both Amazon Linux AMI and
Amazon Linux 2 so the redhatrelease detector correctly detects Amazon
Linux-based images.

Fixes #866