clair: add authorization checking
This commit adds authorization via two methods: PSK and a Quay
keyserver.

The PSK simply uses a key specified in the configuration file.

The keyserver authorization checking relies on the extra "kid" JWT
header to look up a public key and validate the claims on the request.

The tests for the keyserver integration are opportunistic and require
passing a "keyserver" flag to the cmd/clair test for initial setup. Once
initial setup is done, the key and server information will be cached and
used opportunistically on future test runs.