Path routing, Bug Stomping
Traefik is now included in the local dev to confirm
path routing works and to provide a single url for local development.

Cliarctl is configured to point to the traefik instance doing the host
routing by default.

Found a few bugs along the way:
http.Transport in config.HttpClient was not checked for nil and provided
a default so we were getting nil deref panics on test.

HttpTransport client code checked bool before err and was masking an
issue. Actual error was bad local dev config value for indexer_addr in
matcher config

Signed-off-by: ldelossa <ldelossa@redhat.com>