notifier: test mode implementation
this commit enables the notifier to be started in a test mode.
the test mode is enabled via an env var "NOTIFIER_TEST_MODE".
in test mode the notifier will send a notification at each poll
interval.

make targets are provided to start and restart the notifier in test
mode.
the local dev configuration can be configured accordingly to point to
the supported delivery method of your choice.

Signed-off-by: ldelossa <ldelossa@redhat.com>