auth: use better guesses for "aud" claim
This expands the guess at an allowable JWT audience, instead of always
using the request's URL. This was incorrect if any request proxying was
happening, and it almost certainly was.

The code now examines the "Forwarded" header as described in RFC 7239
(https://tools.ietf.org/html/rfc7239#section-4), the various
non-standard "X-Forwarded-" headers, and the request URL and guesses an
acceptable audience based on the first that exists.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>