config: allow disabling notifier
This commit allows for the notifier to be disabled in combo mode by
removing all configuration stanzas. A log line is printed to notify the
operator, although this shouldn't affect operation, as nothing was
configured to happen in this case.

The behavior of notifier mode is slightly altered: it will now refuse to
start if unconfigured.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>