notifier: do less work
previously the notifer would attempt to create notifications for the
same update operation if no affected manifests were discovered.

with this patch we book keep the update operations we encountered and
short circuit the processing of notifications which do not produce
affected manifests.

a receipt in "delivered" status is written to the database directly for
update operations that do not produce affected manifests. this effectively
removes them from subsequent processing and avoides any delivery
attempts.

Signed-off-by: ldelossa <ldelossa@redhat.com>