Direct-delivery: Fix slices in direct notifier
There was a bug in direct notifier that caused only subset of
notification was sent to client. The first iteration over notification
replaced original notification slice with just first "page". This caused
that only one "page" was sent and rest was ignored.

This commit fixes amqp and stomp direct notification delivery.

Signed-off-by: Ales Raszka <araszka@redhat.com>