notifier: remove public key signature scheme
This commit drops all the implementation and support for the public key
signature scheme that the webhook deliverer used.

This feature was originally intended to implement the jwtproxy key
server protocol
(https://github.com/quay/jwtproxy/tree/master/jwt/keyserver/keyregistry)
for easier interoperation with Quay; previous versions of Clair used
jwtproxy to sign and validate traffic to and from Quay. During the
development of v4, we decided to pull the JWT functionality into Clair
proper to ease configuration, and then in the integration with Quay it
was decided to only use a PSK scheme, leaving this key server as a
vestigial piece of functionality.

This change adds some code documentation noting previously exposed
configuration values, but the standalone documentation already omitted
this functionality.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>