auto: add automatic runtime configuration
This change adds an init hook that adjusts GOMAXPROCS, if not set in the
environment already, according to the runtime environment. On Linux,
this means examining CPU quotas.

Doing this will help control latencies by avoiding repeatedly blowing
the CPU budget in containers when run on a very big machine.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>