docker-compose: add paused skopeo container
This is to work around Quay's inability to handle split-view DNS and
port-forwards. Using the API over localhost means clair can't talk to
quay, and using the container name means the host can't talk to quay
over the port forward. The solution I hit on is to just have a waiting
container with skopeo.

The documentation is updated to not recommend spinning up quay by
default, and to mention how to use this odd container arrangement.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>