clair: allow TLS for API server
This change allows Clair to serve its API over TLS (and HTTP/2).
Serving a private API over TLS is generally a pain due to certificate
management and naming issues. However, if required by site policy, it's
now possible without using a sidecar service.

Any reasonably complex Clair deployment will be non-homogeneous, meaning
an application-level load balancer is required. This load balancer will
have to decrypt and re-encrypt the traffic. That is to say, load
balancing based on SNI or TCP connections will not work. As there are
several ways to configure this across several different software stacks,
it is left as an exercise for the reader.

See-also: PROJQUAY-2757
Signed-off-by: Hank Donnay <hdonnay@redhat.com>