auto: fall back to root cgroup
It seems that some container runtimes mount things such that paths
reported from procfs don't line up in sysfs. This falls back to
examining the root cgroup if a specific one doesn't exist; as noted in
the code, if this isn't actually the process' cgroup, it'll be the root
cgroup and end up matching the default GOMAXPROCS guess.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>