matcher: Update matcher client to match server definition
The matcher client is currently creating requests to the matcher
component with a trailing slash, which for the read operations
doesn't match what the server is expecting.
ie. GET /matcher/api/v1/internal/update_operation/. This removes
the trailing slash from both URL paths that the matcher client
is requesting. And adds a dedicated path variable for deleting
update_operations.

Signed-off-by: crozzy <joseph.crosland@gmail.com>