contrib: Wipe all the contents of /tmp on container start
To ensure PVC storage isn't filling up this will wipe /tmp
of any dangling layers that might be hanging out. Also,
remove the init container as it was concerned with doing
the same job but only gets run if the pod is restarted not
if the container runs into trouble.

Signed-off-by: crozzy <joseph.crosland@gmail.com>