config: update module to remove x/sys dependency
Dependabot flagged [GHSA-p782-xgp4-8hr8] in the config module. This is
only used here via the x/tools repo, which is only used by the "tools"
import. Go >=1.17 allows `go run` to specify a module version, so update
the module to require that version and update the `go generate`
directive to use that go run invocation.

[GHSA-p782-xgp4-8hr8]: https://github.com/advisories/GHSA-p782-xgp4-8hr8

Signed-off-by: Hank Donnay <hdonnay@redhat.com>
Signed-off-by: kdeekshithsai7373@gmail.com <Sai.Deekshith.Kotakonda@ibm.com>