cmd: implement modular configs
This takes a page out of k8s' and systemd's playbooks and implements
drop-in configs. The method of handling JSON and YAML is borrowed from
k8s, but simplified because the configurations aren't arbitrary. The
drop-in scheme is inspired by systemd's drop-in overrides.

This allows for operators to modularize the config and put secrets in a
different file than the main config. It also allows for easier
configuration skew -- please note that running multiple different
configurations in the same system is not supported.

Signed-off-by: Hank Donnay <hdonnay@redhat.com>