Documentation: add more information on how to test and get started
In `testing.md`, it was not explicitly stated that the protocol is
needed when specifying the `src` and there was only a podman example
shown for how to copy.

In `getting_started.md` it was not mentioned that the config file has to
be provided when running `clairctl report` and there was no advice on
how to run directly on local image without having to start Quay.

Signed-off-by: gabriel-farache <gfarache@redhat.com>
Signed-off-by: Hank Donnay <hdonnay@redhat.com>