admin: add pre v4.7.3 admin command to create index
In order to facilitate faster deletes we need to add a migration to
add an index in v4.7.3. This change adds an admin command to allow
users to "manually" create the index CONCURRENTLY before the migration
to avoid any down-time. This is something for users with larger
indexer DBs to consider.

Signed-off-by: crozzy <joseph.crosland@gmail.com>