chore: use the merge-multiple directive when downloading binaries
From the docs: When multiple artifacts are matched, this changes the behavior
of the destination directories. If true, the downloaded artifacts will be in
the same directory specified by path. If false, the downloaded artifacts will
be extracted into individual named directories within the specified path.

So essentially, the artifacts were being downloaded into different dirs
and hence the release failure.

Signed-off-by: crozzy <joseph.crosland@gmail.com>